refactor(podcasts): move podcast data and types into a capability module
`components/podcast/` held two kinds of file with different lifetimes: six
legacy UI components, and the type plus data boundary added when the page
was converted to TypeScript. Only the latter survives a redesign, so the
directory said nothing useful about ownership.

Move them to `modules/podcast/` and draw the boundary at that seam:

- `podcast-adapter.ts` -> `modules/podcast/podcast-data.ts`
- `podcast.types.ts` -> `modules/podcast/podcast-types.ts`

Fold the fetch into the module as `fetchPodcastList`, so the page no
longer needs to know the static JSON URL or the transport response shape;
its settled-result handler collapses to passing the list through. Narrow
the module's exports to that one function -- the view-model mapper has no
outside consumer and stays private.

The page keeps what it owns. Pages Router already is the page layer, so
route composition, cache policy, fetch orchestration, and the 404 rule
remain in `pages/podcasts/index.tsx` rather than moving behind a second
page layer that would only add a file to open.

The module stays flat with no `server/` subdirectory. Nothing on this path
is server-only: the schema and mapper are pure, and `fetchStaticJsonByUrl`
is deliberately hybrid, reading the GCS mount on the server and falling
back to HTTP in the browser. A `server/` folder would claim a constraint
the code does not have. The genuinely server-only code in this repo lives
in `utils/server-side-only/` and `firebase/admin`, and is already marked
there.

Behavior is unchanged. Verified that the built `/podcasts` client chunk
still contains none of the schema, monitor, or boundary identifiers.
